refactor: userinput IOmanager 분리
